requirements.txt
flask
Flask-Migrate
python-dotenv
psycopg2-binary
Flask-Seeder
dependency-injector
pydantic